agents/requirements.txt
anthropic==0.39.0
python-dotenv==1.0.0
pydantic==2.10.0
asyncio-contextmanager==1.0.0
